rework gamemanager, games instanciation and how games unfold
rework the gamemanager to just have a reference to all games in the hub and to load or unload a game.
Games are now in a prefab, not anymore in a scene. There is a scriptable object with data for a game and the logic of how the game work is in a strategy object, which is specific for each game logic. A scriptable object is created which contain the logic of the game.

To create a new logic game, you need to create the strategy class for the logic, and a scriptable object factory to easily add the logic to your new game